Fixed a bug for you
Your old code on line 78

  if again == 'yes' or 'y' or 'Y' or 'YES':

Always evaluates to True because you are comparing if again is equal to 'yes' or True, or True, or True 
You need to evaluate each string to again, because otherwise you are evaluating if string is True which is always True.

Also , good job! I've seen that you have been consisting the last few days. And your github almost looks like a New Year resolution challenge to commit atleast a little bit each day.